moved main.js to root dir
The only place Nodejs likes main.js is in the root directory. The public directory is for frontend and static files (HTML,CSS,JS stuff directly related to websites), the views directory is for dynamic/template files that the server will deal with. Anyways it wont work if main.js is in another directory even if it is inline with the package.json file.